+++ front/pages/api/v1/w/[wId]/assistant/conversations/[cId]/mentions/suggestions.test.ts
@@ -1,5 +1,5 @@
 import type { RequestMethod } from "node-mocks-http";
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 
 import { Authenticator } from "@app/lib/auth";
 import { AgentConfigurationFactory } from "@app/tests/utils/AgentConfigurationFactory";
@@ -9,30 +9,12 @@ import {
   createPublicApiMockRequest,
 } from "@app/tests/utils/generic_public_api_tests";
 import { MembershipFactory } from "@app/tests/utils/MembershipFactory";
+import { mockElasticsearch } from "@app/tests/utils/mocks/elasticsearch";
 import { UserFactory } from "@app/tests/utils/UserFactory";
 
 import handler from "./suggestions";
 
-// Mock Elasticsearch
-vi.mock("@app/lib/api/elasticsearch", async (importOriginal) => {
-  const actual: any = await importOriginal();
-  return {
-    ...actual,
-    withEs: vi.fn(async (fn: any) => {
-      const mockClient = {
-        search: vi.fn().mockResolvedValue({
-          hits: { hits: [], total: { value: 0 } },
-        }),
-      };
-      // Mock successful result
-      return {
-        isOk: () => true,
-        isErr: () => false,
-        value: await fn(mockClient),
-      };
-    }),
-  };
-});
+mockElasticsearch();
 
 async function setupTest(method: RequestMethod = "GET") {
   const { req, res, workspace } = await createPublicApiMockRequest({
